+++ lib/jsonable.dart
@@ -1,10 +1,7 @@
 library jsonable;
 
 import 'package:jsonable/src/jsonable.dart';
-import 'package:jsonable/src/validator/validator.dart';
-
 export "./src/jsonable.dart" show Jsonable;
-export './src/validator/validator.dart' show Validator;
 export "./exceptions.dart";
 export "./src/validator/rules.dart" show Rules;
 
@@ -20,7 +17,7 @@ abstract class JType<E> {
   Jsonable parent;
   dynamic keyname;
   E get get => this._value;
-  void set(E value) => this._value = value;
+  void set set(E value) => this._value = value;
 }
 
 abstract class JString extends JType<String>
@@ -32,43 +29,49 @@ abstract class JString extends JType<String>
 }
 
 abstract class JNum extends JType<num> implements Comparable<num> {
-  JNum({num initialValue, Validator validator})
-      : super(
+  JNum({
+    num initialValue,
+  }) : super(
           value: initialValue,
         );
 }
 
 abstract class JBool extends JType<bool> {
-  JBool({bool initialValue, Validator validator})
-      : super(
+  JBool({
+    bool initialValue,
+  }) : super(
           value: initialValue,
         );
 }
 
 abstract class JList<E> extends JType<List<E>> implements List<E> {
-  JList({List<E> initialValue, Validator validator})
-      : super(
+  JList({
+    List<E> initialValue,
+  }) : super(
           value: initialValue,
         );
 }
 
 abstract class JClass<E extends Jsonable> extends JType<E> {
-  JClass({Jsonable initialValue, Validator validator})
-      : super(
+  JClass({
+    Jsonable initialValue,
+  }) : super(
           value: initialValue,
         );
 }
 
 abstract class JMap<E, R> extends JType<Map> implements Map<E, R> {
-  JMap({Map initialValue, Validator validator})
-      : super(
+  JMap({
+    Map initialValue,
+  }) : super(
           value: initialValue,
         );
 }
 
 abstract class JDynamic extends JType<dynamic> {
-  JDynamic({dynamic initialValue, Validator validator})
-      : super(
+  JDynamic({
+    dynamic initialValue,
+  }) : super(
           value: initialValue,
         );
 }